WebNotes/api/requirements.txt
psycopg2-binary
fastapi
uvicorn[standard]
pandas
sqlalchemy